requirements.txt
alembic==1.7.6
anyio==3.4.0
asgiref==3.4.1
black==22.1.0
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
dnspython==2.1.0
email-validator==1.1.3
fastapi==0.70.1
flake8==4.0.1
h11==0.12.0
httptools==0.2.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
Mako==1.1.6
MarkupSafe==2.0.1
mccabe==0.6.1
mypy-extensions==0.4.3
orjson==3.6.5
pathspec==0.9.0
platformdirs==2.5.0
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pydantic==1.9.0
pyflakes==2.4.0
python-dotenv==0.19.2
python-multipart==0.0.5
PyYAML==5.4.1
requests==2.26.0
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.31
starlette==0.16.0
tomli==2.0.1
typing_extensions==4.0.1
ujson==4.3.0
urllib3==1.26.7
uvicorn==0.15.0
uvloop==0.16.0
watchgod==0.7
websockets==10.1
